scanner: Fix over-consuming on !=<> characters
I started out with peeking at the current and next chars but
at some point I thought it would be equally clever to just set
the fact that I saw a two-width char in "in" and do it that way.

Turns out that that breaks simple cases like !), since recording
the BANG token with c=! and i=) will advance to the next character
after the ), skipping it.