parser: Add first brittle pass at recursive descent
Again, to a fault I tried NOT to use set! (which means passing
a lot of state in between the functions...). The ugly bit here
means storing both the overall expression AND the resulting tokens
in the return value (since in the act of parsing an expression,
we may consume an unknown number of tokens).

It seems to work though!

$ ./fisl.scm
> (1 + 1) * 2 != 5
(!= (* (group (+ 1 1)) 2) 5)

Look at that beautiful s-expression :)