repl: Add ,q command and move exit outside run-prompt
I've been toying around with emacs and one of the really cool things
is being able to send updates to chicken REPL. Sometimes I want to
exit the run-prompt loop without ending the scheme process though!

So this patch allows exiting (run-prompt) with ,q (since C-d seems to
be finnicky).