refactor: Drop use of CHICKEN modules (for now)
I've been toying around with Emacs and running an interactive repl
that I can C-x C-e to update functions I just changed (without having
to respawn the full process, etc.

This is a bit easier when not using CHICKEN modules (effectively
exporting everything for interaction with in the repl!).

I may go back at some point (hiding implementation is good), but for
now this is nice!