repl: Add threads and nrepl for live image updates
This was a deep rabbit hole, mostly because (read-line) is blocking I/O
and _all_ threads block when blocking I/O is hit (except tcp). This was
counterintuitive (isn't the whole point of threads to swap when on I/O?)
but nevertheless, I got it to work!

Here's my setup:

1. Start the CHICKEN interpreter via the executable (drops into Lox
   interpreter):

$ ./fisl.scm
> print 1 + 1;
2
>

2. In another tmux split, create a fifo and connect to the 1234 nrepl
   port (credit: http://www.foldling.org/hacking.html#2012-08-05):

$ mkfifo repl
$ cat > repl & nc localhost 1234 < repl

3. Lastly, in Vim, run the following to send the current paragraph to
   the repl on C-c C-c:

:nnoremap <C-c><C-c> vap:w >> repl<CR>

---

Demo:

* Go to (prompt), update s/display "> "/display ">> "/
* C-c C-c
* Hit enter in fisl -- notice prompt is changed!