assignment: Fix assignment vs definition bug
I had intermingled assignment and definition such that

var x = 5;

and

x = 5;

did the same thing, when really the first should create a new binding
in the current environment and the second should update the value in
*whichever environment has x defined (the closest one)*.

This was found while implementing 'while' since the following was an
infinite loop:

var x = 10;
while (x > 0) {
    print x;
    x = x - 1;
}

With the splitting of var -> env-def! and = -> env-set! in this patch,
it works as expected!